Version 0.2.0
- Moved icon files into the .jar (in the streamassist package).  You
don't have to locate an icon directory anymore.  As a result, configs
from 0.1 won't work anymore (but you can always make another).
- Added a set of melee icons
- Added + and - buttons to score fields.  These increment and decrement
the score by 1 respectively.  If the value in the score fields isn't a
number, it will reset the value to 0.  Score field cannot be decremented
below 0 (but it can be manually set to a negative number if you really
want).  Clicking the + and - buttons updates the score live on the
stream (but not icons).
- Added a swap players button that swaps the scores, names, and icons of
both players.
- Added a description field to give a match description (e.g. displaying
"Winner's Finals" above the match).
- Made internal changes to the way icons are updated.  If an icon was
not changed, the program will not update that icon.