Completed Checkbox Addition
- Checking an unchecked box now asks for a file.  If no file is selected
and there was no file originally associated with that checkbox, the
fields associated with that checkbox will be disabled.  If no file is
selected but a file was previously associated with that field, the
previously used file will be used.  If a file is selected, it will
replace the current file.